warningsss

+++ MonkeyDelivery/Src/Control/UI/Sliders/SliderBase.cpp
@@ -1,7 +1,7 @@
 #include "SliderBase.h"
 #include <iostream>
 
-SliderBase::SliderBase(Game* game, int x, int y):GameObject(game){
+SliderBase::SliderBase(Game* game, double x, double y):GameObject(game){
 	setTexture(UI_label);
 	setPosition(x, y);
 }
