+++ MonkeyDelivery/Src/Items/Flashlight.cpp
@@ -19,7 +19,7 @@ bool Flashlight::useObject() {
 			cout << " Hola linternita";
 			player_->changeFLState(true);
 			player_->changeLState(false);
-			sdlutils().soundEffects().at("lantern").setVolume(game_->getSoundEfectsVolume() * game_->getGeneralVolume());
+			sdlutils().soundEffects().at("lantern").setVolume((int)(game_->getSoundEfectsVolume() * game_->getGeneralVolume()));
 			sdlutils().soundEffects().at("lantern").play(0, 1);
 
 		}
@@ -27,7 +27,7 @@ bool Flashlight::useObject() {
 			cout << "Adios linternita";
 			player_->changeFLState(false);
 			player_->setOrientation("off");
-			sdlutils().soundEffects().at("lantern").setVolume(game_->getSoundEfectsVolume() * game_->getGeneralVolume());
+			sdlutils().soundEffects().at("lantern").setVolume((int)(game_->getSoundEfectsVolume() * game_->getGeneralVolume()));
 			sdlutils().soundEffects().at("lantern").play(0, 1);
 
 		}
