+++ MonkeyDelivery/Src/Items/Repel.cpp
@@ -13,7 +13,7 @@ void Repel::setCorrectTexture(){
 
 bool Repel::useObject(){
     player_->initPowerUp(repelente);
-    sdlutils().soundEffects().at("soda").setVolume(game_->getSoundEfectsVolume() * game_->getGeneralVolume());
+    sdlutils().soundEffects().at("soda").setVolume((int)(game_->getSoundEfectsVolume() * game_->getGeneralVolume()));
     sdlutils().soundEffects().at("soda").play(0, 1);
     return true;
 }
